# Automatic matrixworld calculation