--- a/editor/js/Viewport.js
+++ b/editor/js/Viewport.js
@@ -619,9 +619,6 @@ var Viewport = function ( editor ) {
 
 	function render() {
 
-		sceneHelpers.updateMatrixWorld();
-		scene.updateMatrixWorld();
-
 		renderer.clear();
 		renderer.render( scene, camera );