--- a/examples/canvas_interactive_particles.html
+++ b/examples/canvas_interactive_particles.html
@@ -150,7 +150,6 @@
 				camera.position.z = radius * Math.cos( THREE.Math.degToRad( theta ) );
 				camera.lookAt( scene.position );
 
-				camera.updateMatrixWorld();
 
 				// find intersections