--- a/examples/js/WaterShader.js
+++ b/examples/js/WaterShader.js
@@ -203,9 +203,6 @@ THREE.Water.prototype.updateTextureMatrix = function () {
 
 	}
 
-	this.updateMatrixWorld();
-	this.camera.updateMatrixWorld();
-
 	this.mirrorWorldPosition.setFromMatrixPosition( this.matrixWorld );
 	this.cameraWorldPosition.setFromMatrixPosition( this.camera.matrixWorld );
 
@@ -238,7 +235,6 @@ THREE.Water.prototype.updateTextureMatrix = function () {
 	this.mirrorCamera.aspect = this.camera.aspect;
 
 	this.mirrorCamera.updateProjectionMatrix();
-	this.mirrorCamera.updateMatrixWorld();
 	this.mirrorCamera.matrixWorldInverse.getInverse( this.mirrorCamera.matrixWorld );
 
 	// Update the texture matrix