--- a/examples/js/effects/VREffect.js
+++ b/examples/js/effects/VREffect.js
@@ -133,8 +133,6 @@ THREE.VREffect = function ( renderer, onError ) {
 			renderer.enableScissorTest( true );
 			renderer.clear();
 
-			if ( camera.parent === null ) camera.updateMatrixWorld();
-
 			cameraL.projectionMatrix = fovToProjection( eyeFOVL, true, camera.near, camera.far );
 			cameraR.projectionMatrix = fovToProjection( eyeFOVR, true, camera.near, camera.far );