--- a/examples/js/renderers/CSS2DRenderer.js
+++ b/examples/js/renderers/CSS2DRenderer.js
@@ -88,7 +88,7 @@ THREE.CSS2DRenderer = function () {
 
 	this.render = function ( scene, camera ) {
 
-		scene.updateMatrixWorld();
+		scene.updateChildrenMatrixWorld();
 
 		if ( camera.parent === null ) camera.updateMatrixWorld();