--- a/examples/js/renderers/RaytracingRenderer.js
+++ b/examples/js/renderers/RaytracingRenderer.js
@@ -489,7 +489,7 @@ THREE.RaytracingRenderer = function ( parameters ) {
 
 		// update scene graph
 
-		if ( scene.autoUpdate === true ) scene.updateMatrixWorld();
+		if ( scene.autoUpdate === true ) scene.updateChildrenMatrixWorld();
 
 		// update camera matrices