--- a/examples/webgl_interactive_raycasting_pointcloud.html
+++ b/examples/webgl_interactive_raycasting_pointcloud.html
@@ -319,7 +319,6 @@
 			function render() {
 
 				camera.applyMatrix( rotateY );
-				camera.updateMatrixWorld();
 
 				raycaster.setFromCamera( mouse, camera );