--- a/src/extras/helpers/DirectionalLightHelper.js
+++ b/src/extras/helpers/DirectionalLightHelper.js
@@ -9,7 +9,6 @@ THREE.DirectionalLightHelper = function ( light, size ) {
 	THREE.Object3D.call( this );
 
 	this.light = light;
-	this.light.updateMatrixWorld();
 
 	this.matrix = light.matrixWorld;
 	this.matrixAutoUpdate = false;