--- a/src/extras/helpers/HemisphereLightHelper.js
+++ b/src/extras/helpers/HemisphereLightHelper.js
@@ -8,7 +8,6 @@ THREE.HemisphereLightHelper = function ( light, sphereSize ) {
 	THREE.Object3D.call( this );
 
 	this.light = light;
-	this.light.updateMatrixWorld();
 
 	this.matrix = light.matrixWorld;
 	this.matrixAutoUpdate = false;