Continuation of #6851

This is the solution of updating matrixworld without the need to call updatematrixWorld().

What has changed:

- Properties:
    - `Matrix` and `matrixWorld` are always up to date (They get calculated when called). 
    - `Matrix` can be set to a specific matrix. `matrixAutoUpdate` should be set to false.

- Methods:
    - `updateMatrix`: checks if position, rotation and scale haven't changed and recalculates only when changed Before this were always recalculated. This should only be called when matrixAutoUpdate is false and the matrix is in need of an update.
    - `_updateMatrixWorld`: This is an internal method. This calculates the matrixworld based on the previous calculated parent matrixworld and matrix. This is also the method that needs to be overridden when the matrixworld needs to be calculated differently. [example](https://github.com/gero3/three.js/blob/MatrixworldUpdate/examples/js/Gyroscope.js#L14-L63)
    - `updateMatrixWorld`: This calculates the matrixworld by calculating the parents matrixWorld and its onw matrix. Internally it calls `_updateMatrixWorld`.
    - `_updateChildrenMatrixWorld`: This is an internal method. This calculates the matrixworld based on the previous calculated parent matrixworld and matrix and off all its children.  Internally it calls `_updateMatrixWorld`.
    - `updateChildrenMatrixWorld`:This calculates the matrixworld by calculating the parents matrixWorld and its own matrix and that of its descendants. Internally it calls `_updateChildrenMatrixWorld`.

@WestLangley, @mrdoob