@WestLangley 2015-08-19
So users should never have to call `updateMatrix()` or `updateMatrixWorld()` or `updateChildrenMatrixWorld()`. Correct?

Also, the bug that `updateMatrixWorld()` did not update parents is corrected by this PR.

In the renderers, you call `scene.updateChildrenMatrixWorld()` each render loop. Why, exactly, is this beneficial? In other words, why is it faster this way?

And of course, many thanks. : - )