@gero3 2015-08-19
> So users should never have to call updateMatrix() or updateMatrixWorld() or updateChildrenMatrixWorld(). Correct?

correct

> Also, the bug that updateMatrixWorld() did not update parents is corrected by this PR.

indeed
 
> In the renderers, you call scene.updateChildrenMatrixWorld() each render loop. Why, exactly, is this beneficial? In other words, why is it faster this way?

This is needed because of deep hierarchies and the renderers should make use of cached matrices.

> And of course, many thanks. : - )