@gero3 2015-08-19
This has the same performance as current dev.