@gero3 2015-08-23
Is there still a problem with this PR??