@mrdoob 2015-08-28
Sorry for the delay. Having a look now.