@mrdoob 2015-09-02
Merging this at the moment 😉