@mrdoob 2015-09-02
Unfortunately the performance regression is a bit too much 😕

Chrome:
```
webgl_performance.html: 56fps to 53fps
webgl_performance_static.html: 60fps to 53fps
webgl_performance_deep.html: 56fps to 51fps
```

Firefox:
```
webgl_performance.html: 48fps to 40fps
webgl_performance_static.html: 48fps to 45fps
webgl_performance_deep.html: 55fps to 35fps
```