@gero3 2015-09-07
The above commits should fix the performance problems.