@mrdoob 2015-09-08
I'm really hesitant to merge this one... I think I would be less scared if it was split into different parts.

1. Split `updateMatrixWorld()` to `updateMatrixWorld()`, `updateAncestorsMatrixWorld()`  and `updateChildrenMatrixWorld()`.
2. Replace `this.updateMatrixWorld( true )` and `this.matrixWorld` instances with `this.updateAncestorsMatrixWorld()` or `this.getUpdatedMatrixWorld()` (instead of "magical" getters/setters)

I think the idea of making it just work is great, but I think it would make things extra difficult to spot bugs or performance issues in people's code.