@WestLangley 2015-09-12
In an effort to find a compromise so we can move on, perhaps we can omit the lazy updating of matrices, and just create the necessary methods to do what we need.

1. `updateMatrixWorld()` must produce the correct answer. That means it must bubble up the tree. It doesn't; it traverses down the tree. @gero3 has a solution for that. I propose that `updateMatrixWorld()` be modified to do what its name says it does. It should not update the matrices of children.

2. Since the renderer needs to traverse the scene graph, create another method called `updateMatrixWorldOfChildren()`, or just add a recursive flag: `updateMatrixWorld( recursive )`.